move hooks to one directory
